+++ services/web/server/src/simcore_service_webserver/activity/settings.py
@@ -1,5 +1,3 @@
-from typing import Optional
-
 from aiohttp import web
 from settings_library.prometheus import PrometheusSettings
 
@@ -7,7 +5,7 @@ from .._constants import APP_SETTINGS_KEY
 
 
 def get_plugin_settings(app: web.Application) -> PrometheusSettings:
-    settings: Optional[PrometheusSettings] = app[APP_SETTINGS_KEY].WEBSERVER_ACTIVITY
+    settings: PrometheusSettings | None = app[APP_SETTINGS_KEY].WEBSERVER_ACTIVITY
     assert settings, "setup_settings not called?"  # nosec
     assert isinstance(settings, PrometheusSettings)  # nosec
     return settings
